% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ecoregions}
\alias{ecoregions}
\title{Polygon with terrestrial ecoregion}
\format{An object of class \code{SpatialPolygonsDataFrame} with 1194 rows and 2 columns.}
\usage{
ecoregions
}
\description{
This is a spatially simplified shapefile with terrestrial polygon ecoregion from a portion of South America from (Olson et al. 2001).
}
\examples{
require(ENMTML)
data("ecoregions")
plot(ecoregions, col=rainbow(92)[as.factor(ecoregions@data$ECO_ID)])
}
\references{
Olson, D. M., Dinerstein, E., Wikramanayake, E. D., Burgess, N. D., Powell, G. V., Underwood, E. C., … others. (2001). Terrestrial ecoregions of the world: A new map of life on earth. BioScience, 51(11), 933–938.
}
\keyword{datasets}
