% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENM_TheMetaLandALFA.R
\name{ENMs_TheMetaLand}
\alias{ENMs_TheMetaLand}
\title{Create and process Ecological Niche and Species Distribution Models}
\usage{
ENMs_TheMetaLand(pred_dir, sp, x, y, min_occ = 10, thin_occ, colin_var,
  imp_var, transfer, eval_occ, sp_accessible_area, pres_abs_ratio = 1,
  pseudoabs_method, part, save_part = "N", save_final = "Y", algorithm,
  thr, msdm, ensemble, s_sdm)
}
\arguments{
\item{pred_dir}{character. Directory path with predictors (file formats supported are: ASC, BILL, TIFF or TXT)}

\item{sp}{character. Name of the column with information about species names}

\item{x}{character. Name of the column with information about longitude}

\item{y}{character. Name of the column with information about latitude}

\item{min_occ}{interger. Minimum number of unique occurrences (species with less than this number will be excluded)}

\item{thin_occ}{chracater. Perform a spatial filtering (Thinning) on the presences? (Y/N)}

\item{colin_var}{character. Method to reduce variable collinearity: 
\itemize {
  \item N: Use original variables.
  \item Pearson: Select variables by Pearson correlation (threshold specified by user).
  \item VIF: Variance Inflation Factor (Chatterjee and Hadi 2006).
  \item PCA: Perform a Principal Component Analysis on predictors and use Principal Componets as environmental variables
}}

\item{imp_var}{character. Perform importance of variable and curves response for selected algorithms? (Y/N)}

\item{transfer}{character. Project the model onto another region or time period? (Y/N)}

\item{eval_occ}{character. Use a pre-determined set of occurrences for validation? (Y/N)}

\item{sp_accessible_area}{character. Restrict for each species the accessible area,i.e. the area used to construct the model? (Y/N)}

\item{pres_abs_ratio}{numeric. Presence-Absence ratio (values between 0 and 1)}

\item{pseudoabs_method}{character. Pseudo-absence allocation method:
\itemize{ 
\item RND: Random allocation throughout area used to fit models. 
\item ENV_CONST: Pseudo-absence are environmentally constrained to region with lower suitability values predicted by a Bioclim model. 
\item GEO_CONST: Pseudo-absence are constrained allocated far from occurrences by a geographical buffer. 
}}

\item{part}{character. Partition method for model's validation:
\itemize{
  \item BOOT: Random bootstrap partition (e.g. 70"%" training and 30"%" test).
  \item KFOLD: Random partition in k-fold cross validation.
  \item BAND: Geographic partition structured as bands (latitudinal or longitudinal).
  \item BLOCK: Geographic partition structured as a checkerboard.
}}

\item{save_part}{character. Save .tif files of the partitions ? (Y/N) (Default="Y")}

\item{save_final}{character. Save .tif files of the final model (fitted with all the data)? (Y/N) (Default="Y")}

\item{algorithm}{character. Algorithm to construct ecological niche models: 
\itemize{
  \item MAH: Mahalanobis
  \item DOM: Domain
  \item ENF: Ecological Niche Factor Analysis
  \item MXS: Maxent Simple (only linear and quadratic features, based on MaxNet package)
  \item MXD: Maxent Default (all features, based on MaxNet package)
  \item SVM: Support Vector Machine
  \item GLM: Generalized Linear Model
  \item GAM: Generalizes Additive Model
  \item BRT: Boosted Regression Tree
  \item RDF: Random Forest
  \item MLK: Maximum Likelihood
  \item GAU: Gaussian
}}

\item{thr}{character. Threshold used for presence-absence maps:
\itemize{
  \item LPT: The highest threshold at which there is no omission
  \item MAX_TSS: Threshold at which the sum of the sensitivity and specificity is highest
  \item MAX_KAPPA: the threshold at which kappa is highest ("max kappa")
  \item SENSITIVITY: fixed (specified) sensitivity
  \item JACCARD: the threshold at which Jaccard is highest
  \item SORENSEN: the threshold at which Sorensen is highest
  }}

\item{msdm}{character. Include spatial restrictions. These methods restrict Ecological Niche Models in order to have less potential prediction and turn ENMs closer to species distribution models (SDMs). They are classified in a Priori and a Posteriori methods:

a Priori methods: 

\itemize{
  \item N: Do not perform MSDM
  \item XY: Create two layers latitude and longitude layer (added as a predictor)
  \item MIN: Create a layer with information of the distance from each cell to the closest occurrence (added as a predictor)
  \item CML: Create a layer with information of the summed distance from each cell to ALL occurrences (added as a predictor)
  \item KER: Create a layer with a Gaussian-Kernel on the occurrence data (added as a predictor)
  }
a Posteriori methods
\itemize{
  \item POST: Posterior M-SDM Methods (If chosen, preferred method will be asked later) [NOT added as a predictor]
  \item OBR: Occurrence based restriction, uses the distance between points to exclude far suitable patches (Mendes et al, in prep)
  \item LR: Lower Quantile, select the nearest 25% patches (Mendes et al, in prep)
  \item PRES: Select only the patches with confirmed occurrence data (Mendes et al, in prep)
  \item MCP: Excludes suitable cells outside the Minimum Convex Polygon of the occurrence data (Kremen et al, 2008)
  \item MCP-B: Creates a Buffer around the MCP (distance defined by user; Kremen et al, 2008)
  }}

\item{ensemble}{character. Method used to ensemble different algorithms:
\itemize{
\item N: No ensemble
\item MEAN: Simple average of the different models
\item W_MEAN: Weighted Average
\item SUP: Average of the best models (TSS over the average)
\item PCA: Performs a Principal Component Analysis (PCA) and returns the first axis
\item PCA_SUP: PCA of the best models (TSS over the average)
\item PCA_THR: PCA only with cells above the threshold
}}

\item{s_sdm}{character. Perform a stacked of Species Distribution Model (richness map)? (Y/N)}
}
\description{
Create and process Ecological Niche and Species Distribution Models
}
\examples{
library(ENMTheMetaLand)



}
\references{
\itemize{
\item Kremen, C., Cameron, A., Moilanen, A., Phillips, S. J., Thomas, C. D.,
Beentje, H., . Zjhra, M. L. (2008). Aligning Conservation Priorities Across
Taxa in Madagascar with High-Resolution Planning Tools. Science, 320(5873),
222-226. doi:10.1126/science.1155193
}
}
