% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMTML.R
\name{ENMTML}
\alias{ENMTML}
\title{Create and process Ecological Niche and Species Distribution Models}
\usage{
ENMTML(pred_dir, proj_dir = NULL, occ_file, sp, x, y, min_occ = 10,
  thin_occ = NULL, colin_var = NULL, imp_var = NULL,
  eval_occ = NULL, sp_accessible_area = NULL, pres_abs_ratio = 1,
  pseudoabs_method, part, save_part = "N", save_final = "Y", algorithm,
  thr, msdm, ensemble, cores = 1, s_sdm)
}
\arguments{
\item{pred_dir}{character. Directory path with predictors (file formats supported are: ASC, BILL, TIFF or TXT)}

\item{proj_dir}{character. Directory path containing folders with predictors for different regions or time periods used to project models (file formats supported are: ASC, BILL, TIFF or TXT).}

\item{occ_file}{character. Directory path with tab-delimited TXT file with species names, latitude and longitude}

\item{sp}{character. Name of the column with information about species names}

\item{x}{character. Name of the column with information about longitude}

\item{y}{character. Name of the column with information about latitude}

\item{min_occ}{integer. Minimum number of unique occurrences (species with less than this number will be excluded)}

\item{thin_occ}{character Default NULL. Perform spatial filtering (Thinning) on the presences. For this augment it is necessary provide a vector in which its elements need to have the names 'method' or 'method' and 'distance' (more information below). Three thinning methods are available:
\itemize{
  \item 1-Distance defined by Moran Variogram, usage thin_occ=c(method='1').
  \item 2-Distance defined by 2x cellsize (Haversine Transformation), usage thin_occ=c(method='2').
  \item 3-User defined distance. For this option it is neede provide a vector with two values. Usage thin_occ=c(metho='3', ditance='300'). The second numeric value refers to the distance in km that will be used for thinning. So distance=300 means that all records within a radius of 300 km will be deleted
}}

\item{colin_var}{character. Method to reduce variable collinearity:
\itemize{
  \item PCA: Perform a Principal Component Analysis on predictors and use Principal Components as environmental variables, usage colin_var=c(method='PCA').
  \item VIF: Variance Inflation Factor (Chatterjee and Hadi 2006), usage colin_var=c(method='VIF').
  \item PEARSON: Select variables by Pearson correlation, a threshold of maximum correlation must be specified by user, usage colin_var=c(method='PCA', threshold='0.7').
}}

\item{imp_var}{character. Perform variable importance and curves response for selected algorithms? (Y/N)}

\item{sp_accessible_area}{character. Restrict for each species the accessible area, i.e., the area used to construct the model. It is necessary to provide a vector for this argument. Three methods were implemented
\itemize{
  \item BUFFER and based on species occurrences distance, i.e. EXPLAIN MOOORE. Usage sp_accessible_area=c(method='BUFFER', type='1').
  \item BUFFER and based on A single buffer for all species expressed in km. EXPLAIN MOOORE. Usage sp_accessible_area=c(method='BUFFER', type='2', width='300').
  \item MASK: this method consist in delimit the area used to model calibration based on the polygon where a species occurrences fall. For instance it is possible delimit the calibration area based on ecorregion shapefile. For this option it is necessary inform the path to the file that will be used as mask. Next file format can be loaded '.bil', '.asc', '.tif', '.shp', and '.txt'. Usage sp_accessible_area=c(method='MASK', filepath='C:/Users/mycomputer/ecoregion/olson.shp').
}}

\item{pres_abs_ratio}{numeric. Presence-Absence ratio (values between 0 and 1)}

\item{pseudoabs_method}{character. Pseudo-absence allocation method. It is necessary to provide a vector for this argument. Only one method can be chosen. The next methods are implemented:
\itemize{
\item RND: Random allocation throughout area used to fit models. Usage pseudoabs_method=c(method='RND').
\item ENV_CONST: Pseudo-absences are environmentally constrained to a region with lower suitability values predicted by a Bioclim model. Usage pseudoabs_method=c(method='ENV_CONST').
\item GEO_CONST: Pseudo-absences are allocated far from occurrences based on a geographical buffer. For this method it is necessary provie a second value wich express the buffer width in km. Usage pseudoabs_method=c(method='GEO_CONST', width='50').
\item GEO_ENV_CONST: Pseudo-absences are constrained environmentally (based on Bioclim model) but distributed geographically far from occurrences based on a geographical buffer. For this method it is necessary provie a second value wich express the buffer width in km. Usage pseudoabs_method=c(method='GEO_ENV_CONST', width='50')
\item GEO_ENV_KM_CONST: Pseudo-absences are constrained on a three-level procedure; it is similar to the GEO_ENV_CONST with an additional step which distributes the pseudo-absences in the environmental space using k-means cluster analysis. For this method it is necessary provie a second value wich express the buffer width in km. Usage pseudoabs_method=c(method='GEO_ENV_KM_CONST', width='50')
}}

\item{part}{character. Partition method for model's validation. Only one method can be chosen. It is necessary to provide a vector for this argument. The next methods are implemented:
\itemize{
  \item BOOT: Random bootstrap partition (e.g. 70 % training and 30 % test). Usage •    part=c(method='BOOT', replicates='2',  proportion='0.7'). 'replicate' refers to the number of replicates, it assumes a value >=1. 'proportion' refres to the proportion of occurrences used for fitting the model, it assumes a value >0 and <=1.
  \item KFOLD: Random partition in k-fold cross-validation. Usage part=c(method= 'KFOLD', folds='5'). 'folds' referes to the number of k-folds for patitioning, it assume value >=1.
  \item BAND: Geographic partition structured as bands (latitudinal(1) or longitudinal(2)). Usage part=c(method= 'BLOCK', type='1'). 'type' refers to the bands disposition
  \item BLOCK: Geographic partition structured as a checkerboard. Usage part=c(method= 'BLOCK').
}}

\item{save_part}{character. Save .tif files of the partitions? (Y/N) (Default="Y")}

\item{save_final}{character. Save .tif files of the final model (fitted with all the data)? (Y/N) (Default="Y")}

\item{algorithm}{character. Algorithm to construct ecological niche models (it is possible to use more than one method):
\itemize{
  \item BIO: Bioclim
  \item MAH: Mahalanobis
  \item DOM: Domain
  \item ENF: Ecological Niche Factor Analysis
  \item MXS: Maxent Simple (only linear and quadratic features, based on MaxNet package)
  \item MXD: Maxent Default (all features, based on MaxNet package)
  \item SVM: Support Vector Machine
  \item GLM: Generalized Linear Model
  \item GAM: Generalizes Additive Model
  \item BRT: Boosted Regression Tree
  \item RDF: Random Forest
  \item MLK: Maximum Likelihood
  \item GAU: Gaussian Process
}}

\item{thr}{character. Threshold used for presence-absence predictions. It is possible to use more than one threshol type. It is necessary to provide a vector for this argument:
\itemize{
  \item LPT: The highest threshold at which there is no omission. Usage thr=c(type='LPT').
  \item MAX_TSS: Threshold at which the sum of the sensitivity and specificity is highest. Usage thr=c(type='MAX_TSS').
  \item MAX_KAPPA: The threshold at which kappa is highest ("max kappa"). Usage thr=c(type='MAX_KAPPA').
  \item SENSITIVITY: Fixed (specified) sensitivity. For this type of threshold thr must be use as thr=c(type='SENSITIVITY', sens='0.6'). 'sens' refers to models will be bynarized using this suitability value. Note that this method assume 'sens' value for all algorithm.
  \item JACCARD: The threshold at which Jaccard is highest. Usage thr=c(type='JACCARD').
  \item SORENSEN: The threshold at which Sorensen is highest. Usage thr=c(type='SORENSEN').
  }
In the case of use more than  thr=c(type=c('LPT', 'MAX_TSS', 'JACCARD')). In the case of SENSITIVITY threshold is used it is necessayr specify the desired sensitivity value, e.g. thr=c(type=c('LPT', 'MAX_TSS', 'SENSITIVITY'), sens='0.8')}

\item{msdm}{character. Include spatial restrictions. These methods restrict Ecological Niche Models in order to have less potential prediction and turn ENMs closer to species distribution models (SDMs). They are classified in a Priori and a Posteriori methods:

a Priori methods:

\itemize{
  \item N: Do not perform MSDM
  \item XY: Create two layers latitude and longitude layer (added as a predictor)
  \item MIN: Create a layer with information of the distance from each cell to the closest occurrence (added as a predictor)
  \item CML: Create a layer with information of the summed distance from each cell to ALL occurrences (added as a predictor)
  \item KER: Create a layer with a Gaussian-Kernel on the occurrence data (added as a predictor)
  }
a Posteriori methods
\itemize{
  \item POST: Posterior M-SDM Methods (If chosen, the preferred method will be asked later)
  \item OBR: Occurrence based restriction, uses the distance between points to exclude far suitable patches (Mendes et al., in prep)
  \item LR: Lower Quantile, select the nearest 25% patches (Mendes et al., in prep)
  \item PRES: Select only the patches with confirmed occurrence data (Mendes et al, in prep)
  \item MCP: Excludes suitable cells outside the Minimum Convex Polygon of the occurrence data (Kremen et al., 2008)
  \item MCP-B: Creates a Buffer around the MCP (distance defined by user; Kremen et al., 2008)
  }}

\item{ensemble}{character. Method used to ensemble different algorithms (it is possible to use more than one method):
\itemize{
\item N: No ensemble
\item MEAN: Simple average of the different models
\item W_MEAN: Weighted Average
\item SUP: Average of the best models (TSS over the average)
\item PCA: Performs a Principal Component Analysis (PCA) and returns the first axis
\item PCA_SUP: PCA of the best models (TSS over the average)
\item PCA_THR: PCA only with cells above the threshold
}}

\item{cores}{numeric. Define the number number of CPU cores to run modeling procedures in parallel.}

\item{s_sdm}{character. Perform a stacked of Species Distribution Model (richness map)? (Y/N)}
}
\description{
Create and process Ecological Niche and Species Distribution Models
}
\examples{
library(ENMTML)


}
\references{
\itemize{
\item Kremen, C., Cameron, A., Moilanen, A., Phillips, S. J., Thomas, C. D.,
Beentje, H., . Zjhra, M. L. (2008). Aligning Conservation Priorities Across
Taxa in Madagascar with High-Resolution Planning Tools. Science, 320(5873),
222-226. doi:10.1126/science.1155193
}
}
