% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{env}
\alias{env}
\title{Raster with 5 bioclimatic variables for current climate conditions}
\format{
An object of class \code{RasterBrick} of dimension 246 x 282 x 5.
}
\usage{
env
}
\description{
A RasterBrick with Bioclimatic variables (Bio1, Bio3, Bio4, Bio12 and Bio15) from a portion of South America (source Worldclim: https://worldclim.org/version2).
}
\examples{
require(ENMTML)
require(raster)
data("env")
predictors
plot(env)
}
\keyword{datasets}
