% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{m}
\alias{m}
\title{Area used to fit the models of each species}
\format{A list of SpatialPolygons with the area use to fitting the ENMs (each element belongs to a virtual species)}
\usage{
m
}
\description{
Area used to fit the models of each species
}
\examples{

require(raster)
data("m")
data("occ")
data("predictors")
names(m)

par(mfrow=c(2,3))
for(i in unique(occ[,1])){
plot(predictors[[1]], col='gray', legend=FALSE, main=i)
p <- occ \%>\% dplyr::filter(species==i)
points(p[,2:3], col='red', pch=19)
plot(m[[i]], add=T)
}
par(mfrow=c(1,1))
}
\keyword{datasets}
