% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{predictors}
\alias{predictors}
\title{Raster with 19 bioclimatic varibles}
\format{A RasterStack with environmental variables in a portion of South America}
\usage{
predictors
}
\description{
Raster with 19 bioclimatic varibles
}
\examples{

require(raster)
data("predictors")]
predictors
plot(predictors)
}
\keyword{datasets}
