% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{occ}
\alias{occ}
\title{Virtual species occurrences.}
\format{A data.frame with 215 rows and 3 variables:
\describe{
  \item{sp}{virtual species names}
  \item{x}{longitude of species occurrences}
  \item{y}{latitude of species occurrences}
}}
\usage{
occ
}
\description{
A dataset containing occurrences of five virtual species throughout South America.
}
\examples{
require(ENMTML)
data("predictors")
data("occ")

plot(predictors[[1]], col='gray')
points(points(occ[,2:3],
             col=rainbow(5)[as.factor(occ$species)],
             pch=19))
}
\keyword{datasets}
